Basic Sketch Tool
Here is the basic sketch that I exported as an application for my laptop. It can sketch, clear, undo, hide the key commands and start new lines which is my current solution for handling how the shapes try to close each other hiding previous lines. I have a comment in on a potential intensive fix however it may be easier to just store each point in an array and make every shape go between to points. I plan to add a colour picker and slider for depth of the pen in future.